services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
celery[redis]==5.4.0
redis==5.0.6
python-multipart==0.0.9
pydantic==2.7.3
pydantic-settings==2.3.1
# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.3.0
httpx==0.27.0
# Storage
supabase==2.5.0
